refactored code to make it more readable. hopefully. Added callbacks.h/cpp for all callback routines
